Abstract out personal best modal into NewPersonalBestModal react component
