+++ src/components/Speed/PersonalBests/NewPersonalBestModal.css
@@ -1,0 +1,3 @@
+.modalInput {
+  margin: 1rem 0;
+}
