+++ src/components/Speed/PersonalBests/PersonalBests.js
@@ -1,32 +1,13 @@
 import React from 'react';
 import { withFirebase } from '../../../Firebase/index';
 import { JumpalButton } from '../../CustomComponents/core';
+import NewPersonalBestModal from './NewPersonalBestModal';
 import ReactLoading from 'react-loading';
-import Button from '@mui/material/Button';
 import Modal from '@mui/material/Modal';
-import DateTime from 'react-datetime';
-import Select from 'react-select';
 
 import { StyledHeaderTableCell, StyledTableCell, StyledTableRow, StyledTableContainer } from '../../CustomComponents/table';
 import Table from '@mui/material/Table';
 import TableRow from '@mui/material/TableRow';
-
-import Box from '@mui/material/Box';
-import Typography from '@mui/material/Typography';
-
-import { styles } from '../../CustomComponents/constants';
-
-const options = [
-  { value: '1x30sec Running Step', label: '1x30sec Running Step' },
-  { value: '1x60sec Running Step', label: '1x60sec Running Step' },
-  { value: '1x30sec Double Unders', label: '1x30sec Double Unders' },
-  { value: '1x60sec Double Unders', label: '1x60sec Double Unders' },
-  { value: '1x180sec Running Step', label: '1x180sec Running Step' },
-  { value: '1x240sec Running Step', label: '1x240sec Running Step' },
-  { value: 'Consecutive Triple Unders', label: 'Consecutive Triple Unders' },
-  { value: '2x30sec Double Unders', label: '2x30sec Double Unders' },
-  { value: '4x30sec Speed Relay', label: '4x30sec Speed Relay' },
-];
 
 class PersonalBests extends React.Component {
   constructor(props) {
@@ -49,16 +30,9 @@ class PersonalBests extends React.Component {
     this.renderAllData = this.renderAllData.bind(this);
     this.renderTableHeader = this.renderTableHeader.bind(this);
     this.onPersonalBestsUpdate = this.onPersonalBestsUpdate.bind(this);
-    this.toggleNewPersonalBest = this.toggleNewPersonalBest.bind(this);
     this.handleDelete = this.handleDelete.bind(this);
 
-    // New personal best modal methods
-    this.saveNewPersonalBest = this.saveNewPersonalBest.bind(this);
-    this.handleEventChange = this.handleEventChange.bind(this);
-    this.handleScoreChange = this.handleScoreChange.bind(this);
-    this.handleTimeChange = this.handleTimeChange.bind(this);
-    this.timeStamp = this.timeStamp.bind(this);
-
+    // Attach event listener to personal best data of current user
     this.ref = this.props.firebase.db.ref('users')
       .child(this.props.firebase.auth.currentUser.uid)
       .child('personal-bests');
@@ -92,15 +66,8 @@ class PersonalBests extends React.Component {
   }
 
   componentWillUnmount() {
-    // detach all listeners to this reference when component unmounts
+    // detach listeners personal best of current user when component unmounts
     this.ref.off();
-  }
-
-  toggleNewPersonalBest() {
-    this.setState({
-      openNewPersonalBest: !this.state.openNewPersonalBest,
-      newPersonalBestTime: new Date(),
-    })
   }
 
   onPersonalBestsUpdate(snapshot) {
@@ -121,109 +88,6 @@ class PersonalBests extends React.Component {
       .child(event)
       .remove()
     }
-  }
-
-  // New personal best modal methods
-  handleEventChange(event) {
-    this.setState({ 
-      newPersonalBestEvent: event,
-      newPersonalBestEventColor: '#383838',
-     });
-
-  }
-
-  handleScoreChange(event) {
-    this.setState({ newPersonalBestScore: event.target.value });
-  }
-
-  handleTimeChange(time) {
-    this.setState({ newPersonalBestTime: time });
-  } 
-
-  timeStamp(time) {
-    time = new Date(time)
-    // Create an array with the current month, day and time
-    let date = [ time.getMonth() + 1, time.getDate(), time.getFullYear() ];
-
-    // Return the formatted string
-    return date.join("/");
-  }
-
-  saveNewPersonalBest(event) {
-    event.preventDefault();
-    this.props.firebase.db.ref('users')
-    .child(this.props.firebase.auth.currentUser.uid)
-    .child('personal-bests')
-    .child(this.state.newPersonalBestEvent.value)
-    .set({
-      score: this.state.newPersonalBestScore,
-      time: this.timeStamp(this.state.newPersonalBestTime),
-    });
-    
-    window.alert('New Personal Best saved successfully!');
-    this.toggleNewPersonalBest();
-  }
-
-  renderNewPersonalBestModal() {
-    return (
-      <>
-        <div className="jumpalCenteredButton">
-          <JumpalButton onClick={this.toggleNewPersonalBest}>
-            Add New Personal Best
-          </JumpalButton>
-        </div>
-  
-        <Modal
-          open={this.state.openNewPersonalBest}
-          onClose={this.toggleNewPersonalBest}
-          aria-labelledby="modal-modal-title"
-          aria-describedby="modal-modal-description"
-        >
-          <Box sx={styles.modalStyle}>
-            <Typography id="modal-modal-title" variant="h6" component="h2">
-              New Personal Best
-            </Typography>
-            <form className="jumpalForm" ref={(el) => this.PbFormRef = el}>
-              <div>
-                {/* Time Input */}
-                <label>Time: 
-                  <br />
-                  <DateTime
-                    onChange={this.handleTimeChange} 
-                    value={this.state.newPersonalBestTime}
-                    timeFormat={false}
-                  />
-                </label>
-                {/* Event Input */}
-                <label>Event: 
-                  <Select  
-                    value={this.state.newPersonalBestEvent} 
-                    onChange={this.handleEventChange} 
-                    options={options}
-                  />
-                </label>
-
-                {/* Score Input */}
-                <label>
-                  Score:
-                  <input 
-                    className="jumpalInput" 
-                    type="number" 
-                    min="0" 
-                    placeholder="Enter your speed score" 
-                    onChange={this.handleScoreChange}>
-                    </input>
-                </label>
-              </div>
-            </form>
-            {this.state.user 
-                ? <Button variant="success" onClick={this.saveNewPersonalBest}>Save</Button>
-                : <Button variant="success" onClick={this.saveNewPersonalBest} disabled>Save</Button>
-            }
-          </Box>
-        </Modal>
-      </>
-    );
   }
 
   renderTableHeader() {
@@ -256,6 +120,7 @@ class PersonalBests extends React.Component {
       )
     });
   }
+
   render() {
     if (this.state.isDataLoaded) {
       if (this.state.personalBests && this.state.personalBests.length !== 0 && this.state.personalBests[0]){
@@ -263,7 +128,7 @@ class PersonalBests extends React.Component {
         
         return (
           <div className="componentContentDiv">
-            {this.renderNewPersonalBestModal()}
+            <NewPersonalBestModal />
             <h2>My Personal Bests</h2>
             <StyledTableContainer>
               <Table>
@@ -276,7 +141,7 @@ class PersonalBests extends React.Component {
       } else {
         return (
           <div>
-            {this.renderNewPersonalBestModal()}
+            <NewPersonalBestModal />
             <h2>My Personal Bests</h2>
               <p className="loading">Start by entering a new personal best record above.</p>
           </div>
